[refactor] hooks 및 types 패키지 리팩토링 진행 (#73)
* refactor: useJwtStore의 패키지 이동

* refactor: useSelectedCharacterStore의 패키지 이동

* refactor: ApiError의 패키지 이동

* refactor: JwtError의 패키지 이동

* refactor: RuntimeError의 패키지 이동

* chore: 미사용파일(types/zustand.ts) 삭제

* refactor: API_METHOD_TYPE, API_STATUS의 패키지 이동

* refactor: RootStackParamList의 패키지 이동

* refactor: constant -> config로 패키지 네이밍 변경

* fix: jwt.ts 내의 import 변경
