+++ fourtoon-cookie/src/apis/character.ts
@@ -1,5 +1,5 @@
-import { API_METHOD_TYPE, API_STATUS } from "../constants/api";
-import { ApiError } from "../error/ApiError";
+import { API_METHOD_TYPE, API_STATUS } from "../types/api";
+import { ApiError } from "../types/error/ApiError";
 import type { Character } from "../types/character";
 import { requestApi } from "./api";
