+++ fourtoon-cookie/src/config/appinfo.ts
