+++ fourtoon-cookie/src/config/diary.ts
