+++ fourtoon-cookie/src/pages/CharacterSelectPage/Header/Header.tsx
@@ -2,7 +2,7 @@ import { Text, View } from "react-native";
 import BackButton from "../../../components/common/BackButton/BackButton";
 import handleError from "../../../error/errorhandler";
 import { GlobalErrorInfoType } from "../../../types/error";
-import { useSelectedCharacterStore } from "../../../store/selectedCharacter";
+import { useSelectedCharacterStore } from "../../../hooks/store/selectedCharacter";
 
 import * as S from "./Header.styled";
