+++ fourtoon-cookie/src/pages/DiaryWritePage/DiaryWritePage.tsx
@@ -5,15 +5,15 @@ import { NativeStackScreenProps } from "@react-navigation/native-stack";
 import Header from "./Header/Header";
 import TextInputLayout from "./TextInputLayout/TextInputLayout";
 
-import { RootStackParamList } from "../../constants/routing";
+import { RootStackParamList } from "../../types/routing";
 
 import * as S from "./DiaryWritePage.styled";
 import { GlobalErrorInfoType } from "../../types/error";
 import { LocalDate } from "@js-joda/core";
 
-import { RuntimeError } from "../../error/RuntimeError";
+import { RuntimeError } from "../../types/error/RuntimeError";
 import handleError from "../../error/errorhandler";
-import { useSelectedCharacterStore } from "../../store/selectedCharacter";
+import { useSelectedCharacterStore } from "../../hooks/store/selectedCharacter";
 import WriteDoneButtonLayout from "./WriteDoneButtonLayout/WriteDoneButtonLayout";
 import { NavigationProp, useNavigation } from "@react-navigation/native";
 import { useDiaryById } from "../../hooks/server/diary";
