+++ fourtoon-cookie/src/types/error/RuntimeError.ts
