--- fourtoon-cookie/src/types/zustand.ts
@@ -1,3 +1,0 @@
-
-
-type setType = (partial: unknown, replace?: boolean | undefined) => void;
